Use int64 where applicable
In Go, the actual size of "int" is based on your platform, so in order to be more explicit, use int64 where it will be needed.

+++ Euler010/Euler010.go
@@ -4,7 +4,7 @@ func main() {
 	sum := 0
 
 	for i := 1; i < 2000000; i++ {
-		if isPrime(i) {
+		if isPrime(int64(i)) {
 			sum += i
 		}
 	}
@@ -12,8 +12,8 @@ func main() {
 	println(sum)
 }
 
-func isPrime(n int) bool {
-	for i := 2; i*i <= n; i++ {
+func isPrime(n int64) bool {
+	for i := int64(2); i*i <= n; i++ {
 		if n%i == 0 {
 			return false
 		}
